requirements.txt
streamlit>=1.28.0
python-toon>=0.1.0
pandas>=2.0.0